% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-biblio.R
\name{align_biblio_columns}
\alias{align_biblio_columns}
\title{Align columns before row-binding bibliographic files}
\usage{
align_biblio_columns(dfs)
}
\description{
Align columns before row-binding bibliographic files
}
\keyword{internal}
