% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-biblio.R
\name{read_generic}
\alias{read_generic}
\title{Read a generic CSV with user-specified columns}
\usage{
read_generic(
  file,
  id = NULL,
  sep = ";",
  authors = NULL,
  keywords = NULL,
  references = NULL,
  countries = NULL,
  affiliations = NULL,
  journal = NULL,
  list_cols = NULL
)
}
\description{
Read a generic CSV with user-specified columns
}
\keyword{internal}
