% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-biblio.R
\name{read_single_biblio}
\alias{read_single_biblio}
\title{Read a single bibliometric file}
\usage{
read_single_biblio(
  file,
  format,
  id,
  sep,
  authors = NULL,
  keywords = NULL,
  references = NULL,
  countries = NULL,
  affiliations = NULL,
  journal = NULL,
  list_cols = NULL,
  ...
)
}
\description{
Read a single bibliometric file
}
\keyword{internal}
